First commit of HTM balancer. This projects attempts to balance an inverted pendulum using HTM neural network. This version uses a very simplified pendulum at the moment and simply rewards cells when the angle is closer to vertical.
The top rows of cells are the angle measurements. The middle cells represent
close to 90 degrees (vertical) orientation. The second set of rows represents
the feedback command from higher levels. The bottom set of rows is the
angle command outputs for that level.

Only level 0 commands are actually used to control the simulated inverted
pendulum.